docs: split agent instructions by package
