+++ mcp_plex/common/__init__.py
@@ -1,0 +1,7 @@
+"""Shared utilities for the loader and server packages."""
+
+from __future__ import annotations
+
+from .cache import MediaCache
+
+__all__ = ["MediaCache"]
