+++ mcp_plex/server/__main__.py
@@ -1,0 +1,8 @@
+"""CLI entry point for ``python -m mcp_plex.server``."""
+
+from __future__ import annotations
+
+from . import main
+
+if __name__ == "__main__":
+    main()
